+++ src/collective/contact/core/tests/test_behaviors.py
@@ -1,5 +1,5 @@
 # -*- coding: utf8 -*-
-import unittest2 as unittest
+import unittest
 
 from zope.component import getUtility
 from zope.event import notify