+++ src/collective/contact/core/tests/test_content_types.py
@@ -1,6 +1,6 @@
 # -*- coding: utf8 -*-
 
-import unittest2 as unittest
+import unittest
 
 import datetime