+++ src/collective/contact/core/tests/test_search.py
@@ -1,7 +1,7 @@
 # -*- coding: utf8 -*-
 
 import datetime
-import unittest2 as unittest
+import unittest
 
 from plone import api